Add fuzzer for libffi library
This fuzzer for the libffi library may be run on-device or host.
Note that as libffi is very arch-specific, fuzzing on host will
provide different coverage than on-device fuzzing.

This is a patch & resubmit of c260aa78a9bce30a642637ccaf61409907f1c254,
due to an arch-specific build issue & revert.

Test: Tested on pixel 3a, with approximately 57% coverage (arm64) and
11k executions/sec after a short run with no initially provided corpus.

Signed-off-by: Michael Ensing <michael.ensing@leviathansecurity.com>
Change-Id: I999d97388289d9f43cb2f28fe7eaf6d9e2051515