Adding a tool to enable quick testing of attestation
The attached README provides instructions for building and running this
testing tool. The intent of this tool is to provide a way for OEMs to
quickly diagnose any major issues that might occur with respect to
attestaiton if a device isn't properly provisioned on the factory line,
or there's a bug in how the root of trust is being reported.

The source is used to put together an APK which will be pushed to device
and generate an attestation. The APK will check the relevant fields in
the certificate and log any errors to logcat, which will be read back
out to the terminal by the python script which executes the test process.

This initial commit only supports Keymaster 4.0 (e.g. attestation
version 3). 4.1 and 3.0 coverage will be added.

Fixes: 154879902
Test: reference the README
Change-Id: I9a60b2e6c5820857143ed39a2234df2be13c8bcf