A gdb pretty-printer that print variables in json format
This is a pretty-printer that print variables in json format. Its
requirement doc is at go/timeless-fuzz-reports-pretty-printer-reqs. read
README.md for ways to run it.
Test: the gdb library cannot be imported unless loading the script in
gdb. So there's no automatic testcase for this script. However, several
test examples are provided in test_examples so that user can test in gdb
with these examples manually.

Change-Id: I3738827d5d25bc4ec8aba3c1ee2e11c7c7ed7aae