Always instantiate a usable public key
As well as checking the details in the COSE structure, pass the public
key bytes to the crypto library for full validation of the key. This
will confirm that the key itself is valid, e.g. the EC point is on the
curve.

Previously, the validity of the key is only checked when it is used for
verification and there was no guarantee that the leaf key would ever be
validated.

Test: atest libcert_request_validator_tests
Change-Id: I8a4cad20a54931c2766df0e2064cd4792573297a