Use types and traits to handle CoseKey parameters
The parameters have a label which often have values the come from IANA
constants. Use the `Label` type and `EnumI64` trait in favour of raw
`i64` and casting.

Move the CoseKey paramter access helpers next to the code that uses
then, introduce some constants and imports to make things a bit less
verbose.

Test: atest libcert_request_validator_tests
Change-Id: I3798bc205f8ce5ccf5e8522632d0a4489dcd4c82