Add a cxx bridge for libcert_request_validator
Add an FFI bridge to allow the libkeymint_remote_prov_support C++
library, used by VTS tests and the factory extraction tool, to call the
libcert_request_validator Rust library.

The Rust details are hidden within the bridge library and a simple C++
interface is exposed that currently just does what is needed to allow
the existing code to migrate.

Bug: 254510672
Test: `atest VtsHalRemotelyProvisionedComponentTargetTest` on cf
Change-Id: I0e3fa333858ee9cc08a8158ae2a7741dd9777bff