Expose more from hwtrust for DICE chain validation
There are a lot of upcoming requirements (see below) that place
restrictions on the CSR and the DICE chain. This is a first step to
make some of the infrastructure necessary to utilize
the already-existing functionality inside hwtrust.

Bug: 372843164
Bug: 376297620

A lot of parsing happens in `remote_prov_utils`. It would be nice to
have all parsing and validation occur in hwtrust.

Test: atest libhwtrust_tests
Change-Id: Ie73e98b84d4fc95bc9992f3eb1746b77e4d48b90